test(reader): split shell construction coverage
Why:
- Continue shrinking the app-host XCTest bundle by moving package-owned reader shell construction coverage into the BibleUI package tests.
- Keep hamburger drawer and speak mini-player construction covered without app launch or app delegate dependencies.

What Changed:
- Added ReaderShellConstructionTests under the BibleUI package test target.
- Moved speak mini-player, navigation drawer, and side-drawer overlay construction tests out of AndBibleTests+AppAndReader.
- Removed app-host imports that only supported the moved shell tests.
- Updated docs/test-architecture-migration-plan.md to reflect the remaining AppAndReader scope.

Validation:
- git diff --cached --check
- python3 scripts/check_repo_standards.py docblocks --all-files
- python3 scripts/check_repo_standards.py source-guards
- GitNexus impact on each moved test reported LOW risk with no callers or affected flows.
- GitNexus detect_changes scope=staged reported no mapped production-flow changes.
- xcodebuild test -project AndBible.xcodeproj -scheme BibleUITests -destination 'platform=iOS Simulator,name=iPhone 17' -derivedDataPath /tmp/andbible-reader-shell-bibleui-dd CODE_SIGNING_ALLOWED=NO -only-testing:BibleUITests/ReaderShellConstructionTests
- xcodebuild test -quiet -project AndBible.xcodeproj -scheme AndBible -destination 'platform=iOS Simulator,name=iPhone 17' -derivedDataPath /tmp/andbible-reader-shell-app-dd CODE_SIGNING_ALLOWED=NO -only-testing:AndBibleTests/AndBibleTests/testPassageChooserUsesFullScreenChooserShell -only-testing:AndBibleTests/AndBibleTests/testReaderPassageChooserOverlayDoesNotExposeDeadCancellationAPI -only-testing:AndBibleTests/AndBibleTests/testReaderPassageChooserCallSitesUseSharedWorkspaceTitleSource -only-testing:AndBibleTests/AndBibleTests/testPassageChooserDrawerDoesNotDependOnNativeNavigationBarInsideReader -only-testing:AndBibleTests/AndBibleTests/testReaderPassageChooserProgressProvidersCaptureSnapshotsOnce -only-testing:AndBibleTests/AndBibleTests/testApplicationDelegateSceneConfigurationUsesWindowSceneDelegate -only-testing:AndBibleTests/AndBibleTests/testColorARGBByteClampsIntermediatePickerComponents -only-testing:AndBibleTests/AndBibleTests/testColorARGBIntClampsOutOfRangeComponents
- Adversarial sub-agent review found no staged-diff issues.

Impact:
- No production behavior changes.
- Reduces app-host test scope while preserving package-owned BibleUI shell coverage.
- Leaves passage chooser source/progress and color helper tests for later migration slices.
